a-for-loop.py
Life gets hard for everyone and as a programmer,
the basics when solving a problem get fuzzy. 

Don't forget FOR loops. Because they can make
a difference in your code and in your life too

P.S. I added this code block as my LinkedIn background image.
Only those who discover it will understand it (muahaha)